backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
muselsl>=2.2.2
pylsl>=1.16.2
redis>=5.2.0
scipy>=1.14.0
scikit-learn>=1.6.0
joblib>=1.4.0
google-generativeai>=0.8.0
numpy>=1.26.0
pygame>=2.6.0
python-dotenv>=1.0.0
pydantic>=2.10.0
